oiijoj

+++ API/User.py
@@ -4,8 +4,8 @@ class User():
                  service_dog : bool, hard_hearing : bool,
                  visually_impaired : bool):
         self.name = name
-        self.level = xp / 1000
-        self.xp = xp % 1000,
+        self.level = int(xp / 1000)
+        self.xp = int(xp % 1000),
         self.wheelchair = wheelchair
         self.service_dog = service_dog
         self.hard_hearing = hard_hearing
